samples/molecule_agent/requirements.txt
pydantic-settings>=2.8.1
google-adk
python-dotenv
google-cloud-aiplatform
google-genai
